backend/requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pybit==5.6.2
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.0
apscheduler==3.10.4
websockets==12.0
httpx==0.26.0
